Support millisecond precision in event timestamps
Changed TimeUnixSeconds fields from long to decimal in all event models to allow millisecond precision. Updated NetworkStateUpdater to convert decimal Unix timestamps to DateTime with millisecond accuracy. Added comprehensive tests to verify deserialization, validation, and conversion of both integer and decimal timestamps for backward compatibility.

+++ node-api/Models/L2Trace.cs
@@ -17,7 +17,7 @@ public record L2Trace : NetworkEventDatagram
     public required string ReportFrom { get; init; }
 
     [JsonPropertyName("time")]
-    public long? TimeUnixSeconds { get; init; }
+    public decimal? TimeUnixSeconds { get; init; }
 
     /// <summary>
     /// Port ID, e.g. "3" or "4mlink" (Required)
