+++ node-api/Models/NodeDownEvent.cs
@@ -15,7 +15,7 @@ public record NodeDownEvent : NetworkEventDatagram
     /// Timestamp (secs since 1/1/70) (Optional)
     /// </summary>
     [JsonPropertyName("time")]
-    public long? TimeUnixSeconds { get; init; }
+    public decimal? TimeUnixSeconds { get; init; }
 
     /// <summary>
     /// Node Callsign (Required)
